Geoview-geochart
Panel has a new callback for when the panel has completed its opening transition
Reworked panel opened handle
Update geochart.tsx
Panels synch
Some configs
Improved createCodeSnippetUsingIDs to be able to point to more than 1 script and stack them in the code snippet
Better templates and showing of code snippets
Cleaned up the generation of default values via parsing, the default values are coming from the Chart itself.
Now handling an error when the Chart fails to render
Changed the package.json to point to the #develop branch instead of main
Cleanup and support for theme
Just some template values update
GeoChart not caring for the redrawing state timer anymore (simplified)
Now with layers for map1 for an attempt to attach to queries done event
Added dependencies to useEffect()
Now parsing new useSteps parameter
Handling the error and using the ChartValidator static parsing for the errors
Merged the data and options parameters for the plugin into 1 single json
Schema stuff
Geochart configuration progress
Added a rule about having an await if a function has the 'async' keyword and enforcing the return to be a Promise.
Changed from an error to a warning
Fixed a the params comments on getGeoviewLayerByIdAsync
New event to load the config for the plugin dynamically (extra)
Cleaner event payloads
Deactivated the sliders events (temporarily)
Now loading the config in json 'live' in the template
Commit to not lose track. Lots of work done, going deeper..

From the Component, I have no themes setup, so when I import the UI.Select component from cgpv there are no theme.
I've tried to create a 'cgpv theme' over in my component, but I then needed a ThemeProvider component to wrap my application, which I didn't find as a export in cgpv.
So I added support to the Select component when no themes are set for now.

I couldn't find out how to make the MenuItemProps.content property work when passed via TypeMenuItemProps. So I'm doing it more explicitly and this works fine.

Simplified the component injection in checkboxlist-enhanced to follow pattern implemented in Select

Rush update

Re-added the hyper link to the demo which disappeared at some time