New TypeFeatureInfoEntryPartial type (#1443)
New Esri-related common functions
Turned the styling of select-style theme usage optional (because there are no themes when importing a cgpv ui component in an external package where there are no maps and no ThemeProvider - discuss)
Tweaked the checkbox-list-enhanced.tsx to align with a pattern in select.tsx (part of this commit too)
I couldn't find out how to make the MenuItemProps.content property work when passed via TypeMenuItemProps. The Component was failing to unpack. So, I'm directly injecting the component prop - more explicitly - and this works fine.
Exposing more props to the slider-base, like : orientation, valueLabelFormat, getAriaValueText
Added a useEffect to the component so that it listens when the parent component wants to update the content (because the value was using a state internally and would never change ulteriorly)
Added some TODOs

Renaming the inner variable to clearer name
Updated the TODO with information